+++ js/constants.js
@@ -1,0 +1,39 @@
+
+export const PHOTOS_COUNT = 25;
+
+export const MIN_LIKES = 15;
+export const MAX_LIKES = 200;
+
+export const MIN_COMMENTS = 0;
+export const MAX_COMMENTS = 30;
+
+export const COMMENTS_MAX_ID = 1000;
+
+export const MIN_AVATAR = 1;
+export const MAX_AVATAR = 6;
+
+export const NAMES = ['Артём', 'Виктор', 'Владимир', 'Илья', 'Александр', 'Анатолий', 'Сергей'];
+
+export const MESSAGES = [
+  " Всё отлично!",
+  "В целом всё неплохо. Но не всё.",
+  "Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.",
+  "Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.",
+  "Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.",
+  "Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!"
+];
+
+export const DESCRIPTIONS = [
+  'Закат над морем', 'Горы в тумане', 'Цветущий сад', 'Старый замок на холме',
+  'Лесная тропа в осенний день', 'Городские огни ночью', 'Пейзаж с озером и горами',
+  'Детская улыбка на фоне радуги', 'Пляж с белым песком и пальмами', 'Снежные вершины в солнечный день',
+  'Уютный домик в деревне', 'Цветы весной в парке', 'Космическое небо с яркими звездами',
+  'Рыбаки на реке на рассвете', 'Древние руины среди природы',
+  'Современный городской пейзаж', 'Тихий вечер у костра',
+  'Птицы, летающие над морем',
+  'Старые книги на полке библиотеки', 'Водопад в тропическом лесу',
+  'Фотография с высоты птичьего полета города',
+  'Зимний лес, покрытый снегом',
+  'Корабли в гавани на закате', 'Дети, играющие на лужайке',
+  'Танец огня в камине'
+];
